requirements.txt
streamlit==1.36.0
tensorflow==2.20.0
numpy==2.1.0
Pillow==10.4.0
gdown==4.7.1